neotree: icons and file_watcher changes
Change the empty directory icon to be a bit more descriptive and added
some padding to some existing ones to look a bit better in the neotree.
Also, changed the file_watcher behaviour to be enabled by default so
that you don't have to refresh the neotree after each change/commit.